backend/control_center/requirements.txt
tensorflow-cpu==2.12.1
Django==5.1.14
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
django-celery==3.1.17
django-celery-beat==2.7.0
django-cors-headers==4.3.1
django-db-connection-pool[postgresql]==1.2.6
django-environ==0.12.0
django-filter==24.3
django-rest-knox==5.0.2
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-api-key==3.1.0
drf-spectacular==0.28.0
ansible==9.3.0
ansible-core==2.16.4
ansible-runner==2.3.4
docker==7.0.0
gunicorn==20.1.0
getmac==0.9.3
pandas==1.5.3
numpy==1.24.2
psutil==5.9.8
psycopg2==2.9.10
redis==5.0.3
requests==2.28.2
scapy==2.5.0
service-identity==24.1.0
pyOpenSSL==24.1.0
python-dotenv==1.0.0
uvicorn[standard]
pipreqs==0.4.12